feat: bookshop 전체 도서 조회

+++ controller/CartController.js
@@ -1,11 +1,13 @@
 const conn = require(`../mariadb`);
 const { StatusCodes } = require('http-status-codes');
 const jwt = require('jsonwebtoken');
+const ensureAuthorization = require('../auth');
+const jwt = require('jsonwebtoken');
 const dotenv = require('dotenv');
 dotenv.config();
 
 const addToCart = (req, res) => {
-    const { book_id, quantity, user_id } = req.body;
+    const { book_id, quantity } = req.body;
 
     let authorization = ensureAuthorization(req, res);
 
@@ -33,7 +35,6 @@ const addToCart = (req, res) => {
     }
 };
 
-// 장바구니 아이템 목록 조회
 const getCartItems = (req, res) => {
     const { selected } = req.body;
 
@@ -51,9 +52,14 @@ const getCartItems = (req, res) => {
         let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
                 FROM cartItems LEFT JOIN books 
                 ON cartItems.book_id = books.id
-                WHERE user_id=? AND cartItems.id IN (?)`;
+                WHERE user_id=?`;
+        let values = [authorization.id];
+        if (selected) {
+            // 주문서 작성 시 '선택한 장바구니 목록 조회'
+            sql += ` AND cartItems.id IN (?)`;
+            values.push(selected);
+        }
 
-        let values = [authorization.id, selected];
         conn.query(sql, values, (err, results) => {
             if (err) {
                 console.log(err);
@@ -65,37 +71,31 @@ const getCartItems = (req, res) => {
     }
 };
 
-// 장바구니 아이템 삭제
 const removeCartItem = (req, res) => {
-    const cartItemId = req.params.id; // cartItemId
+    let authorization = ensureAuthorization(req, res);
 
-    let sql = 'DELETE FROM cartItems WHERE id = ?';
-    conn.query(sql, cartItemId, (err, results) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message: '로그인 세션이 만료 되었습니다. 다시 로그인 하세요.',
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message: '잘못된 토큰입니다.',
+        });
+    } else {
+        const cartItemId = req.params.id; // cartItemId
 
-        return res.status(StatusCodes.OK).json(results);
-    });
-};
+        let sql = 'DELETE FROM cartItems WHERE id = ?';
+        conn.query(sql, cartItemId, (err, results) => {
+            if (err) {
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
 
-function ensureAuthorization(req, res) {
-    try {
-        let receivedJwt = req.headers['authorization'];
-        console.log('received jwt : ', receivedJwt);
-
-        let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-        console.log(decodedJwt);
-
-        return decodedJwt;
-    } catch (err) {
-        console.log(err.name);
-        console.log(err.message);
-
-        return err;
+            return res.status(StatusCodes.OK).json(results);
+        });
     }
-}
+};
 
 module.exports = {
     addToCart,
